Add Rubocop and run before tests
Rubocop provides a way to ensure we have consistent code formatting
across a project. No more bike-shedding about opinions.

For now I am happy to use the default Rubocop rules. Things do evolve
though, so if you come across a way we format the code and want to
change it, submit a PR with changes to `.rubocop.yml` along with the
reasoning behind the change and see what the team thinks.