Enable GitHub Actions for CI
We want to be able to run a CI server but don't want to manage it ourselves.

GitHub Actions is great because our code is already hosted there. Having the CI server hosted there makes sense. No extra credentials to remember, the results are integrated into the pull requests and it has a simple infrastructure as code model that is easy to maintain.

Initially, we want to run tests on every push. We should also enable the repo level checks that prevent pull requests being merged with failing tests.