Add SimpleCov
We want a way to track our test code coverage.
[SimpleCov](https://github.com/colszowka/simplecov) does this and works
well with RSpec.

Coverage is calculated on every test run and outputs the stats alongside
the test suite results.

For now we are relying on you checking this number when you make your
changes. In the future we can look at automating this check and making
it part of the PR review process using a service like Codacy.