Add the entrypoint for the parsing process
We want to be able to run the parsing script from the command line. This
requires an executable script that accepts a filename as an input.

I opted to start with the basics of starting the script and verifying
that we have an argument present. This gives us a foundation to build on
in future changes.

While doing this, I also set up some logic to determine if this script
is being called from RSpec or from the command line.

This also allows us to write tests free from worrying about running a
sub-process.