Count total page views
As a first step, we want to count the total page views per path.

This makes an assumption that each line contains only a single page
view. If this changes then the object that counts the page views will
need to be modified.

While building this I opted to create separate objects to handle each
step of the process.

`Parser` is a co-ordinating object, holding knowledge of the disparate
parts of the system but not executing any logic of its own.

The tests for `Parser` are effectively integration tests. Ensuring that
data flows between the different objects successfully.

`PageCount` is responsible for counting the page views in the log file.
It only needs to know the definitions of the log format and how to group
those entries when counting.

Finally we have a `Formatter` object. It takes a list of counts and
knows how to convert them into a suitable output format such as a
string.

This would be the point to extend if we want other output formats, eg.
HTML, JSON.

Notice that I chose to use primitives to pass data between objects. This
ensures we don't couple our business logic code to each other. So the
inputs and outputs of `PageCount` and `Formatter` should always be
primitives such as Array, Hash or String rather than instances of
themselves.