Enable new RuboCop cops
There are some new cops that have been added in recent releases of
RuboCop that are not enabled by default.

Given that this is a new project, it is easier for us to enable them now
and ensure we are compatible rather than enable them in the future and
have to potentially handle the changes they introduce.