Add a count type to improve readability of the output
We want to be able to display different types of counts, ie. total or
unique. The current output doesn't have a way of presenting this type.

Modifying `PageCount` to scope the counts by the type and then allowing
the `Formatter` to use this new data structure gives us a simple way to
add any type of count in the future.

This change simply modifies the existing behaviour. It doesn't add any
extra count types.