Count unique page views
We currently only count the total number of page views in the log file
but would also like to know how many unique page views there were.

This makes an assumption that the IP address represents a unique entity
and therefore we count a unique view when the IP address hasn't been
seen before in conjunction with the path.

Building on the work in #10, we are able to add a new type to the Hash
provided to the `Formatter` and this displays our new counts without
needing to modify that.

The count algorithm is simple and based on using a key-value data
structure to efficiently keep track of the number of unique views
in O(n) time rather than the alternative O(n**2) or possibly O(n**3)
approaches that loop through the array multiple times to find the IP
matches.