Relax block and line length checks in specs
There are 2 checks that RuboCop performs that are hard to stick to in
our spec files.

Keeping the block length to 25 lines is a good check for ensuring our
methods don't become overly complex but it becomes a hinderance when
applied to the top level RSpec `describe` block.

Line length is another that can be hard to stick to all the time. The
expressive nature of RSpec tends to quickly hit the 80 character limit.
Most of the time we can use new lines to format the code appropriately
and fix this but there are cases where the expected output is a string
longer than 80 characters. We can't do anything about that other than
ignore the check.